Remove API endpoint and move it into utils
